pyproject.toml
[tool.poetry]
name = "link-shorter"
version = "0.1.0"
description = ""
authors = ["fedor <f.shatruk@thesis.team>"]
readme = "README.md"
packages = [{include = "link_shorter"}]

[tool.poetry.dependencies]
python = "^3.11"
celery = "^5.2.7"
django = "^4.1.7"
django-redis = "^5.2.0"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.5"
serializers = "^0.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
